@@ -13,21 +13,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/cdk": "^5.2.5",
-    "@angular/cli": "1.7.4",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
-    "@angular/material": "^5.2.5",
-    "@angular/material-moment-adapter": "^5.2.5",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
+    "@angular/animations": "6.0.0",
+    "@angular/cdk": "6.0.0",
+    "@angular/cli": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/language-service": "6.0.0",
+    "@angular/material": "6.0.0",
+    "@angular/material-moment-adapter": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "body-parser": "^1.18.2",
     "commander": "^2.15.1",
     "compression": "^1.7.2",
@@ -38,9 +38,11 @@
     "google-spreadsheet": "^2.0.4",
     "moment": "^2.22.1",
     "moment-timezone": "^0.5.16",
-    "rxjs": "^5.0.0",
-    "typescript": ">=2.4.2 <2.7.0",
-    "webpack-dev-middleware": "^3.1.3",
+    "rxjs": "^6.1.0",
+    "typescript": "2.7.2",
     "zone.js": "^0.8.26"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
